Try to stabilize CI by moving specs out of async category
